+++ crates/wdl-lint/src/v1.rs
@@ -10,6 +10,8 @@ mod double_quotes;
 mod ending_newline;
 mod import_placement;
 mod matching_parameter_meta;
+mod missing_metas;
+mod missing_output;
 mod missing_runtime;
 mod no_curly_commands;
 mod pascal_case;
@@ -23,6 +25,8 @@ pub use double_quotes::*;
 pub use ending_newline::*;
 pub use import_placement::*;
 pub use matching_parameter_meta::*;
+pub use missing_metas::*;
+pub use missing_output::*;
 pub use missing_runtime::*;
 pub use no_curly_commands::*;
 pub use pascal_case::*;
@@ -74,6 +78,8 @@ pub fn rules() -> Vec<Box<dyn Rule>> {
         Box::new(CommandSectionMixedIndentationRule),
         Box::new(ImportPlacementRule),
         Box::new(PascalCaseRule),
+        Box::new(MissingMetasRule),
+        Box::new(MissingOutputRule),
     ];
 
     // Ensure all the rule ids are unique and pascal case